+++ docs/changelog.md
@@ -13,6 +13,7 @@ Click on the version in the following list to see all changes since the previous
 ### [Unreleased]
 
 * docs: add Java documentation to model classes
+* feat: import templates from Quick Subtasks for Jira (#74)
 
 ### [23.02.1] - 2023-02-19